[FRONT: S-E/S-VR] Complete establishment detail and reusable reviews

+++ Back/reviews/models.py
@@ -29,7 +29,9 @@ class Review(models.Model):
         blank=True,
     )
 
-    rating = models.PositiveSmallIntegerField(
+    rating = models.DecimalField(
+        max_digits=2,
+        decimal_places=1,
         validators=[
             MinValueValidator(0),
             MaxValueValidator(5),
